Refactoring graph editor out of shapes.rs
