+++ src/rust/lib/graph/src/lib.rs
@@ -1,3 +1,7 @@
+//! NOTE
+//! This file is under a heavy development. It contains commented lines of code and some code may
+//! be of poor quality. Expect drastic changes.
+
 #![feature(associated_type_defaults)]
 #![feature(drain_filter)]
 #![feature(overlapping_marker_traits)]
@@ -15,238 +19,284 @@
 #![warn(missing_copy_implementations)]
 #![warn(missing_debug_implementations)]
 
+
 pub mod component;
-
-use prelude::*;
-
-use ensogl::display;
-use ensogl::display::Sprite;
-use ensogl::display::shape::*;
-use ensogl::display::object::traits::*;
-use ensogl::traits::*;
-use logger::Logger;
-use enso_prelude::std_reexports::fmt::{Formatter, Error};
-use nalgebra::Vector2;
-
-pub mod prelude {
-    pub use ensogl::prelude::*;
-}
-
-pub use component::node;
-pub use component::node::Node;
 
 use ensogl::prelude::*;
 use ensogl::traits::*;
 
 use ensogl::data::color::*;
+use ensogl::display;
+use ensogl::display::Sprite;
 use ensogl::display::navigation::navigator::Navigator;
 use ensogl::display::shape::*;
 use ensogl::display::shape::primitive::system::ShapeSystemDefinition;
 use ensogl::display::shape::Var;
 use ensogl::display::world::*;
 use ensogl::system::web;
+use crate::component::node;
+use crate::component::node::Node;
+use crate::component::node::WeakNode;
+use crate::component::cursor;
+use crate::component::cursor::Cursor;
+use nalgebra::Vector2;
 use shapely::shared;
 use std::any::TypeId;
+use ensogl::control::io::mouse::MouseManager;
+use enso_frp as frp;
+use enso_frp::{frp, Position};
+use enso_frp::Mouse;
+use ensogl::control::io::mouse;
+use enso_frp::core::node::class::EventEmitterPoly;
+use ensogl::display::layout::alignment;
+use ensogl::display::scene;
+use ensogl::display::scene::{Scene, MouseTarget};
+use ensogl::gui::component::StrongRef;
+use ensogl::gui::component::WeakRef;
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
 
-
-// =========================
-// === Library Utilities ===
-// =========================
-
-//pub trait HasSprite {
-//    fn set_sprite(&self, sprite:&Sprite);
-//}
-
-
-// TODO[ao] copied from shapes example, consider where to move it.
-pub mod icons {
-    use super::*;
-
-    pub fn history() -> AnyShape {
-        let radius_diff    = 0.5.px();
-        let corners_radius = 2.0.px();
-        let width_diff     = &corners_radius * 3.0;
-        let offset         = 2.px();
-        let width          = 32.px();
-        let height         = 16.px();
-        let persp_diff1    = 6.px();
-
-        let width2          = &width  - &width_diff;
-        let width3          = &width2 - &width_diff;
-        let corners_radius2 = &corners_radius  - &radius_diff;
-        let corners_radius3 = &corners_radius2 - &radius_diff;
-        let persp_diff2     = &persp_diff1 * 2.0;
-
-        let rect1 = Rect((&width ,&height)).corners_radius(&corners_radius);
-        let rect2 = Rect((&width2,&height)).corners_radius(&corners_radius2).translate_y(&persp_diff1);
-        let rect3 = Rect((&width3,&height)).corners_radius(&corners_radius3).translate_y(&persp_diff2);
-
-        let rect3 = rect3 - rect2.translate_y(&offset);
-        let rect2 = rect2 - rect1.translate_y(&offset);
-
-        let rect1 = rect1.fill(Srgba::new(0.26, 0.69, 0.99, 1.00));
-        let rect2 = rect2.fill(Srgba::new(0.26, 0.69, 0.99, 0.6));
-        let rect3 = rect3.fill(Srgba::new(0.26, 0.69, 0.99, 0.4));
-
-        let icon = (rect3 + rect2 + rect1).translate_y(-persp_diff2/2.0);
-        icon.into()
-    }
-}
-
-fn ring_angle<R,W,A>(inner_radius:R, width:W, angle:A) -> AnyShape
-    where R : Into<Var<Distance<Pixels>>>,
-          W : Into<Var<Distance<Pixels>>>,
-          A : Into<Var<Angle<Radians>>> {
-    let inner_radius = inner_radius.into();
-    let width        = width.into();
-    let angle        = angle.into();
-
-    let angle2  = &angle / 2.0;
-    let radius  = &width / 2.0;
-    let inner   = Circle(&inner_radius);
-    let outer   = Circle(&inner_radius + &width);
-    let section = Plane().cut_angle(&angle);
-    let corner1 = Circle(&radius).translate_y(inner_radius + radius);
-    let corner2 = corner1.rotate(&angle2);
-    let corner1 = corner1.rotate(-&angle2);
-    let ring    = &outer - &inner;
-    let pie     = &ring * &section;
-    let out     = &pie + &corner1 + &corner2;
-//    let out     = out.fill(Srgba::new(0.22,0.83,0.54,1.0));
-//    let out     = out.fill(Srgba::new(0.0,0.0,0.0,0.2));
-    let out     = out.fill(Srgba::new(0.9,0.9,0.9,1.0));
-    out.into()
-}
-
-fn nodes2() -> AnyShape {
-    let node_radius = 32.0;
-    let border_size = 16.0;
-    let node   = Circle(node_radius.px());
-//    let border = Circle((node_radius + border_size).px());
-    let node   = node.fill(Srgb::new(0.97,0.96,0.95));
-//    let node   = node.fill(Srgb::new(0.26,0.69,0.99));
-//    let border = border.fill(Srgba::new(0.0,0.0,0.0,0.06));
-
-    let bg   = Circle((node_radius*2.0).px());
-    let bg   = bg.fill(Srgb::new(0.91,0.91,0.90));
-
-
-//    let shadow1 = Circle((node_radius + border_size).px());
-//    let shadow1_color = LinearGradient::new()
-//        .add(0.0,Srgba::new(0.0,0.0,0.0,0.08).into_linear())
-//        .add(1.0,Srgba::new(0.0,0.0,0.0,0.0).into_linear());
-//    let shadow1_color = SdfSampler::new(shadow1_color).max_distance(border_size).slope(Slope::InvExponent(5.0));
-//    let shadow1       = shadow1.fill(shadow1_color);
-
-    let shadow2 = Circle((node_radius + border_size).px());
-    let shadow2_color = LinearGradient::new()
-        .add(0.0,Srgba::new(0.0,0.0,0.0,0.0).into_linear())
-        .add(1.0,Srgba::new(0.0,0.0,0.0,0.14).into_linear());
-//    let shadow2_color = ExponentSampler::new(shadow2_color);
-    let shadow2_color = SdfSampler::new(shadow2_color).max_distance(border_size).slope(Slope::Exponent(4.0));
-    let shadow2       = shadow2.fill(shadow2_color);
-
-
-    let loader_angle : Var<Angle<Radians>> = "Radians(clamp(input_time/2000.0 - 1.0) * 1.99 * PI)".into();
-    let loader_angle2 = &loader_angle / 2.0;
-    let loader        = ring_angle((node_radius).px(), (border_size).px(), loader_angle);
-    let loader        = loader.rotate(loader_angle2);
-    let loader        = loader.rotate("Radians(input_time/200.0)");
-
-    let icon = icons::history();
-
-
-    let out = loader + shadow2 + node + icon;
-    out.into()
+pub mod prelude {
+    pub use ensogl::prelude::*;
 }
 
 
 
-/// Registers Node shape.
-pub fn register_shapes(world:&World) {
-    todo!()
-//   let node_shape   = nodes2();
-//   let shape_system = ShapeSystemDefinition::new(world,&node_shape);
-//   world.scene().shapes.insert(TypeId::of::<Node>(),shape_system.clone());
+use ensogl::control::event_loop::RawAnimationLoop;
+use ensogl::control::event_loop::AnimationLoop;
+use ensogl::control::event_loop::TimeInfo;
+use ensogl::control::event_loop::FixedFrameRateSampler;
+use ensogl::animation::physics::inertia::DynInertiaSimulator;
+use ensogl::data::OptVec;
+use ensogl::display::object::Id;
+use ensogl::system::web::StyleSetter;
+
+
+#[derive(Clone,CloneRef,Debug,Default)]
+pub struct NodeSet {
+    data : Rc<RefCell<HashMap<Id,Node>>>
 }
 
-
-
-// =============
-// === Graph ===
-// =============
-
-#[derive(Default)]
-pub struct OnEditCallbacks {
-    node_added : Option<Box<dyn Fn(&node::Node) + 'static>>,
-}
-
-impl Debug for OnEditCallbacks {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), Error> {
-        f.write_str("graph::OnEditCallbacks")
-    }
-}
-
-#[derive(Debug,Default)]
-struct GraphData {
-    nodes : Vec<node::Node>,
-}
-
-#[derive(Debug)]
-pub struct Graph {
-    data           : Rc<RefCell<GraphData>>,
-    display_object : display::object::Node,
-    callbacks      : Rc<RefCell<OnEditCallbacks>>,
-    logger         : Logger,
-}
-
-impl Graph {
-    pub fn new(world:&World) -> Self {
-        let logger         = Logger::new("graph");
-        let data           = default();
-        let display_object = display::object::Node::new(&logger);
-        let callbacks      = default();
-        register_shapes(world);
-        Self {data,display_object,callbacks,logger}
+impl NodeSet {
+    pub fn borrow(&self) -> Ref<HashMap<Id,Node>> {
+        self.data.borrow()
     }
 
-    pub fn set_on_edit_callbacks(&self, callbacks: OnEditCallbacks) {
-        *self.callbacks.borrow_mut() = callbacks
+    pub fn take(&self) -> HashMap<Id,Node> {
+        mem::take(&mut *self.data.borrow_mut())
     }
-}
 
-impl<'a> From<&'a Graph> for &'a display::object::Node {
-    fn from(graph: &'a Graph) -> Self {
-        &graph.display_object
+    pub fn insert(&self, node:Node) {
+        self.data.borrow_mut().insert(node.id(),node);
+    }
+
+    pub fn remove(&self, node:&Node) {
+        self.data.borrow_mut().remove(&node.id());
+    }
+
+    pub fn contains(&self, node:&Node) -> bool {
+        self.get(node.id()).is_some()
+    }
+
+    pub fn get(&self, id:Id) -> Option<Node> {
+        self.data.borrow().get(&id).map(|t| t.clone_ref())
     }
 }
 
 
-// === Interface for library users ===
 
-impl Graph {
-    pub fn add_node(&self, new_node:node::Node) {
-        self.display_object.add_child(&new_node);
-        self.data.borrow_mut().nodes.push(new_node);
+#[derive(Clone,CloneRef,Debug,Default)]
+pub struct WeakNodeSet {
+    data : Rc<RefCell<HashMap<Id,WeakNode>>>
+}
+
+impl WeakNodeSet {
+    pub fn borrow(&self) -> Ref<HashMap<Id,WeakNode>> {
+        self.data.borrow()
     }
 
-    pub fn clear_graph(&self) {
-        let mut data = self.data.borrow_mut();
-        for node in &data.nodes {
-            self.display_object.remove_child(node);
+    pub fn take(&self) -> HashMap<Id,WeakNode> {
+        mem::take(&mut *self.data.borrow_mut())
+    }
+
+    pub fn for_each_taken<F:Fn(Node)>(&self,f:F) {
+        self.take().into_iter().for_each(|(_,node)| { node.upgrade().for_each(|n| f(n)) })
+    }
+
+    pub fn insert(&self, node:&Node) {
+        self.data.borrow_mut().insert(node.id(),node.downgrade());
+    }
+
+    pub fn contains(&self, node:&Node) -> bool {
+        self.get(node.id()).is_some()
+    }
+
+    pub fn get(&self, id:Id) -> Option<Node> {
+        self.data.borrow().get(&id).and_then(|t| t.upgrade())
+    }
+}
+
+
+#[derive(Clone,CloneRef,Debug,Default,Shrinkwrap)]
+pub struct WeakNodeSelectionSet {
+    data : WeakNodeSet
+}
+
+impl WeakNodeSelectionSet {
+    pub fn deselect_all(&self) {
+        self.for_each_taken(|node| node.events.deselect.event.emit(()));
+    }
+}
+
+
+
+pub struct Events {
+    pub add_node              : frp::Dynamic<()>,
+    pub remove_selected_nodes : frp::Dynamic<()>,
+}
+
+impl Events {
+    pub fn new() -> Self {
+        frp! {
+            add_node              = source::<()> ();
+            remove_selected_nodes = source::<()> ();
         }
-        data.nodes.clear();
+        Self {add_node,remove_selected_nodes}
     }
 }
 
+pub struct GraphEditor {
+    pub events         : Events,
+    pub selected_nodes : WeakNodeSelectionSet,
+}
 
-// === Interface for GUI events ===
+impl GraphEditor {
 
-impl Graph {
-    pub fn gui_add_node(&self, new_node:node::Node) {
-        self.gui_add_node(new_node.clone());
-        if let Some(callback) = &self.callbacks.borrow().node_added {
-            callback(&new_node)
+    pub fn new(world: &World) -> Self {
+        let scene  = world.scene();
+        let camera = scene.camera();
+        let screen = camera.screen();
+
+
+        let cursor = Cursor::new();
+        web::body().set_style_or_panic("cursor","none");
+
+
+        world.add_child(&cursor);
+
+
+        let events = Events::new();
+    //    web::body().set_style_or_panic("cursor","none");
+
+        let mouse = &scene.mouse.frp;
+
+        let node_set = NodeSet::default();
+
+        let selected_nodes = WeakNodeSelectionSet::default();
+
+        let selected_nodes2 = selected_nodes.clone_ref();
+
+        frp! {
+            mouse_down_position    = mouse.position.sample        (&mouse.on_down);
+            selection_zero         = source::<Position>           ();
+            selection_size_down    = mouse.position.map2          (&mouse_down_position,|m,n|{m-n});
+            selection_size_if_down = selection_size_down.gate     (&mouse.is_down);
+            selection_size_on_down = selection_zero.sample        (&mouse.on_down);
+            selection_size         = selection_size_if_down.merge (&selection_size_on_down);
+
+
+            mouse_down_target      = mouse.on_down.map            (enclose!((scene) move |_| scene.mouse.target.get()));
+
+
+            node_mouse_down = source::<Option<WeakNode>> ();
+
+            foo = events.add_node.map2(&mouse.position, enclose!((node_set,node_mouse_down,world) move |_,pos| {
+                let node = Node::new();
+                let weak_node = node.downgrade();
+                let ttt = node.view.events.mouse_down.map("foo",enclose!((node_mouse_down) move |_| {
+                    node_mouse_down.event.emit(Some(weak_node.clone_ref()))
+                }));
+
+                world.add_child(&node);
+                node.mod_position(|t| {
+                    t.x += pos.x as f32;
+                    t.y += pos.y as f32;
+                });
+
+                node_set.insert(node);
+
+            }));
+
+            bar = events.remove_selected_nodes.map(enclose!((selected_nodes2) move |_| {
+                selected_nodes2.for_each_taken(|node| {
+                    node_set.remove(&node);
+                })
+            }));
+
+            baz = node_mouse_down.map(move |opt_node| {
+                opt_node.for_each_ref(|weak_node| {
+                    weak_node.upgrade().map(|node| {
+                        let is_selected = selected_nodes2.contains(&node);
+                        selected_nodes2.deselect_all();
+                        node.events.select.event.emit(());
+                        selected_nodes2.insert(&node);
+                    })
+                })
+            })
         }
+
+
+        mouse.on_down.map("cursor_press", enclose!((cursor) move |p| {
+            cursor.events.press.event.emit(());
+        }));
+
+        mouse.on_up.map("cursor_release", enclose!((cursor) move |p| {
+            cursor.events.release.event.emit(());
+        }));
+
+        mouse.position.map("cursor_position", enclose!((cursor) move |p| {
+            cursor.set_position(Vector2::new(p.x as f32,p.y as f32));
+        }));
+
+        selection_size.map("cursor_size", enclose!((cursor) move |p| {
+            cursor.set_selection_size(Vector2::new(p.x as f32,p.y as f32));
+        }));
+
+        let selected_nodes2 = selected_nodes.clone_ref();
+        mouse_down_target.map("mouse_down_target", enclose!((scene) move |target| {
+            match target {
+                display::scene::Target::Background => {
+                    selected_nodes2.deselect_all();
+                }
+                display::scene::Target::Symbol {symbol_id, instance_id} => {
+                    scene.shapes.get_mouse_target(&(*instance_id as usize)).for_each(|target| {
+                        target.mouse_down().for_each(|t| t.event.emit(()));
+                    })
+                }
+            }
+        }));
+
+        let add_node_ref = events.add_node.clone_ref();
+        let remove_selected_nodes_ref = events.remove_selected_nodes.clone_ref();
+        let selected_nodes2 = selected_nodes.clone_ref();
+        let world2 = world.clone_ref();
+        let c: Closure<dyn Fn(JsValue)> = Closure::wrap(Box::new(move |val| {
+            let val = val.unchecked_into::<web_sys::KeyboardEvent>();
+            let key = val.key();
+            if      key == "n"         { add_node_ref.event.emit(()) }
+            else if key == "Backspace" {
+                remove_selected_nodes_ref.event.emit(())
+            }
+            else if key == "p" {
+                selected_nodes2.for_each_taken(|node| {
+                    world2.scene().remove_child(&node);
+                })
+            }
+        }));
+        web::document().add_event_listener_with_callback("keydown",c.as_ref().unchecked_ref()).unwrap();
+        c.forget();
+
+
+        Self {events,selected_nodes}
     }
 }
